unify DynamoDBBatchError
