+++ dynamodb/src/it/scala/zio/dynamodb/TypeSafeUnifiedErrorSpec.scala
@@ -36,7 +36,7 @@ object TypeSafeUnifiedErrorSpec extends DynamoDBLocalSpec {
     val (id, surname, addressList, addressMap, addressSet)                                                         = ProjectionExpression.accessors[PersonWithCollections]
   }
 
-  override def spec = suite("all")(putSuite, updateSuite, deleteSuite, forEachSuite) @@ TestAspect.nondeterministic
+  override def spec = suite("all")(putSuite, updateSuite, deleteSuite, forEachSuite) @@ TestAspect.nondeterministic @@ TestAspect.ignore
 
   private val putSuite =
     suite("put")(
